hand to hand on session/requirements.txt
Flask==3.0.0
Pillow==10.0.0
pytesseract==0.3.10
Werkzeug==3.0.0
